Dgrid feedback (#102)
Fix query result source type, and add data to fetch results

The query transform result's source type was an observable
store, but it needs to be a query store because it should
be queryable(since it was queried to create the result in
the first place).

Fetch now has two promise properties, totalLength and
dataLength. For a Store or Storage, these will resolve to
the same value, which is just the total number of items in
the backing storage(which just means the storage itself in
that case). The only reason both are provided on these
interfaces is for consistency.

For a QueryTransformResult, totalLength will still represent
the same value(the total number of items in the underlying
storage). dataLength on the other hand, will resolve to the
number of items that match the QueryTransformResult's
queries.

In all cases, the values of these properties do not take
into account any queries passed to fetch.